+++ tests/contracts/headers/shader_hpp.cpp
@@ -7,3 +7,5 @@
 
 static_assert(!std::is_copy_constructible_v<granit::shader>);
 static_assert(std::is_move_constructible_v<granit::shader>);
+static_assert(std::is_trivially_copyable_v<granit::shader_ref>);
+static_assert(granit::shader_ref::from_native(UINT64_C(7)).native_handle() == UINT64_C(7));